mp4mux: add ftyp and extra atoms to the streamheaders
ftyp and extra atoms are needed in the caps' streamheader
field to reconstruct a valid header from the caps, useful
for fragmented streaming, where the moov is created at the
very beginning with only the useful initialization data.

For fragmented mp4 streams, the header buffers are queued until
the header is completed and they are pushed after setting the
streamheader in caps and before starting to push moof's.
It doesn't change the behaviour with regular or fast_start mp4 files,
where the final moov is only known when finishing the file.

https://bugzilla.gnome.org/show_bug.cgi?id=706509